refactor: update movieId parameter type to Long

+++ repository/src/main/java/com/baghdad/repository/datasource/remote/RemoteMovieDataSource.kt
@@ -4,7 +4,7 @@ import com.baghdad.repository.model.CastMemberDto
 import com.baghdad.repository.model.MovieDto
 
 interface RemoteMovieDataSource {
-    suspend fun getSimilarMovies(movieId: Int): List<MovieDto>
-    suspend fun getMovieDetails(movieId: Int): MovieDto
-    suspend fun getMovieCredits(movieId: Int): List<CastMemberDto>
+    suspend fun getSimilarMovies(movieId: Long): List<MovieDto>
+    suspend fun getMovieDetails(movieId: Long): MovieDto
+    suspend fun getMovieCredits(movieId: Long): List<CastMemberDto>
 }
